vcmi: split CBonusSystemNode, BonusParams and prop
More splitting of HeroBonus.h
