+++ lib/CStack.h
@@ -11,6 +11,7 @@
 #pragma once
 #include "JsonNode.h"
 #include "bonuses/HeroBonus.h"
+#include "bonuses/CBonusSystemNode.h"
 #include "CCreatureHandler.h" //todo: remove
 #include "battle/BattleHex.h"
 #include "mapObjects/CGHeroInstance.h" // for commander serialization
