+++ lib/JsonNode.cpp
@@ -13,8 +13,10 @@
 
 #include "ScopeGuard.h"
 
+#include "bonuses/BonusParams.h"
 #include "bonuses/HeroBonus.h"
 #include "bonuses/Limiters.h"
+#include "bonuses/Propagators.h"
 #include "bonuses/Updaters.h"
 #include "filesystem/Filesystem.h"
 #include "VCMI_Lib.h" //for identifier resolution
