+++ lib/battle/BattleInfo.h
@@ -10,6 +10,7 @@
 #pragma once
 #include "../int3.h"
 #include "../bonuses/HeroBonus.h"
+#include "../bonuses/CBonusSystemNode.h"
 #include "CBattleInfoCallback.h"
 #include "IBattleState.h"
 #include "SiegeInfo.h"
