+++ lib/mapObjects/CArmedInstance.h
@@ -12,6 +12,7 @@
 #include "CObjectHandler.h"
 #include "../CCreatureSet.h"
 #include "../bonuses/CBonusProxy.h"
+#include "../bonuses/CBonusSystemNode.h"
 
 VCMI_LIB_NAMESPACE_BEGIN
